Move `MaybeCurve` to where it is used
