+++ crates/fj-kernel/src/partial/objects/mod.rs
@@ -1,4 +1,3 @@
-pub mod curve;
 pub mod cycle;
 pub mod edge;
 pub mod face;
